Move `get_tensors_partial_shapes` to dev API
