+++ src/core/src/op/util/evaluate_helpers.cpp
@@ -4,8 +4,6 @@
 
 #include "ngraph/op/util/evaluate_helpers.hpp"
 
-#include "openvino/op/util/evaluate_helpers.hpp"
-
 namespace ngraph {
 AxisSet get_normalized_axes_from_tensor(const HostTensorPtr tensor,
                                         const ngraph::Rank& rank,
@@ -17,18 +15,3 @@ AxisSet get_normalized_axes_from_tensor(const HostTensorPtr tensor,
     return AxisSet{normalized_axes};
 }
 }  // namespace ngraph
-
-namespace ov {
-namespace op {
-namespace util {
-std::vector<PartialShape> get_tensors_partial_shapes(const TensorVector& tensors) {
-    std::vector<PartialShape> shapes;
-    shapes.reserve(tensors.size());
-    for (const auto& t : tensors) {
-        shapes.emplace_back(t.get_shape());
-    }
-    return shapes;
-}
-}  // namespace util
-}  // namespace op
-}  // namespace ov
