+++ src/core/src/validation_util.cpp
@@ -1384,5 +1384,14 @@ std::shared_ptr<Constant> get_constant_from_source(const Output<Node>& source) {
         return {};
     }
 }
+
+std::vector<PartialShape> get_tensors_partial_shapes(const TensorVector& tensors) {
+    std::vector<PartialShape> shapes;
+    shapes.reserve(tensors.size());
+    for (const auto& t : tensors) {
+        shapes.emplace_back(t.get_shape());
+    }
+    return shapes;
+}
 }  // namespace util
 }  // namespace ov
